refactor : List<Lotto> 만드는 RandomMachine역할 RandomLottoFactory에 통일
